+++ ASDKSample/ASDKSample/AboutViewController.swift
@@ -27,7 +27,7 @@ class AboutViewController: UIViewController {
         super.viewDidLoad()
 
         var labelStatusText = "Tinkoff Acquiring SDK"
-        if let version = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String, 
+        if let version = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String,
            let build = Bundle.main.object(forInfoDictionaryKey: "CFBundleVersion") as? String {
             labelStatusText.append("\n \(version)(\(build))")
         }